requirements.txt
aiohttp>=3.9.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
tzdata>=2024.1
